ISSUE-57 Added set -x and if exist migrations condition to init_db_and_run.sh

+++ bin/init_db_and_run.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+set -x
 cd marketplace.app
 
 source .venv/bin/activate
@@ -10,7 +11,9 @@ export FLASK_DEBUG=1
 #Перед запуском убедитесь, что у вас запущен postgres в доккере
 #И создана marketplace.db на сервере.
 
-rm -rf migrations
+if [ -d "migrations" ]; then
+  rm -rf migrations
+fi
 flask db init
 flask db migrate
 flask db upgrade
